Change TrackDetailView image
