+++ iMusic.xcodeproj/xcuserdata/wsa-024-23b.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "64E351B3-F037-481A-9DD1-6CF2E96AAFB3"
+   type = "1"
+   version = "2.0">
+</Bucket>
